# docs/swagger.yaml
basePath: /
definitions:
...
  models.Block:
    properties:
      children:
        items:
          $ref: '#/definitions/models.Block'
        type: array
      id:
        type: string
      indent:
        type: integer
      lineNumber:
        type: integer
      name:
        type: string
      parentId:
        type: string
      properties:
        additionalProperties:
          type: string
        type: object
      rawType:
        type: string
      subflowId:
        type: string
      tokens:
        items:
          $ref: '#/definitions/models.BlockToken'
        type: array
      type:
        $ref: '#/definitions/models.BlockType'
      variables:
        items:
          type: string
        type: array
    type: object
...
  models.BlockToken:
    properties:
      target:
        description: The ID or name of the target (subflow name, variable name, etc.)
        type: string
      type:
        description: '"text", "variable", "subflow", "label"'
        type: string
      value:
        description: The text to display
        type: string
    type: object
  models.BlockType:
    enum:
    - ACTION
    - LOOP
    - CONDITION
    - SUBFLOW
    - ERROR_HANDLER
    - COMMENT
    - VARIABLE
    - WAIT
    - ELSE
    - CASE
    - DEFAULT
    - BLOCK
    - SWITCH
    - END
    - UNKNOWN
    type: string
    x-enum-varnames:
    - BlockTypeAction
    - BlockTypeLoop
    - BlockTypeCondition
    - BlockTypeSubflow
    - BlockTypeErrorHandler
    - BlockTypeComment
    - BlockTypeVariable
    - BlockTypeWait
    - BlockTypeElse
    - BlockTypeCase
    - BlockTypeDefault
    - BlockTypeBlock
    - BlockTypeSwitch
    - BlockTypeEnd
    - BlockTypeUnknown
...
  models.FlowDocument:
    properties:
      filePath:
        type: string
      files:
        items:
          $ref: '#/definitions/models.FlowFileInfo'
        type: array
      id:
        type: string
      isFolder:
        type: boolean
      metadata:
        $ref: '#/definitions/models.FlowMetadata'
      name:
        type: string
      parseErrors:
        items:
          $ref: '#/definitions/models.ParseError'
        type: array
      subflows:
        items:
          $ref: '#/definitions/models.Subflow'
        type: array
    type: object
  models.FlowFileInfo:
    properties:
      name:
        type: string
      path:
        type: string
      size:
        type: integer
    type: object
  models.FlowMetadata:
    properties:
      blockCount:
        type: integer
      fileSize:
        type: integer
      maxDepth:
        type: integer
      parsedAt:
        type: string
      rawLineCount:
        type: integer
      subflowCount:
        type: integer
    type: object
...
  models.Highlight:
    properties:
      end:
        type: integer
      start:
        type: integer
    type: object
...
  models.ParseError:
    properties:
      column:
        type: integer
      line:
        type: integer
      message:
        type: string
      severity:
        type: string
      snippet:
        type: string
    type: object
...
  models.RecentFile:
    properties:
      isFolder:
        type: boolean
      lastOpen:
        type: string
      name:
        type: string
      path:
        type: string
      size:
        type: integer
    type: object
...
  models.SearchQuery:
    properties:
      blockTypes:
        items:
          $ref: '#/definitions/models.BlockType'
        type: array
      fuzzy:
        type: boolean
      maxResults:
        type: integer
      text:
        type: string
    type: object
  models.SearchResult:
    properties:
      blockId:
        type: string
      highlights:
        items:
          $ref: '#/definitions/models.Highlight'
        type: array
      matchedField:
        type: string
      matchedText:
        type: string
      score:
        type: integer
      subflowId:
        type: string
    type: object
  models.SearchResults:
    properties:
      durationMs:
        type: integer
      query:
        $ref: '#/definitions/models.SearchQuery'
      results:
        items:
          $ref: '#/definitions/models.SearchResult'
        type: array
      totalCount:
        type: integer
    type: object
...
  models.Subflow:
    properties:
      blocks:
        items:
          $ref: '#/definitions/models.Block'
        type: array
      id:
        type: string
      name:
        type: string
      sourceFile:
        description: file name (without path) this subflow was parsed from
        type: string
      variables:
        items:
          $ref: '#/definitions/models.VariableDecl'
        type: array
    type: object
...
  models.VariableDecl:
    properties:
      initialValue:
        type: string
      name:
        type: string
      scope:
        type: string
      type:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: This is the API for the Pad Analyzer project.
  title: Pad Analyzer API
  version: "1.0"
paths:
...
  /api/flow/clear-recent:
    post:
      description: Clears the entire list of recently opened flow documents.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Clear recent files
      tags:
      - flow
  /api/flow/load-folder:
    post:
      consumes:
      - application/json
      description: Loads a flow document from the specified local folder path. Only
        available in local mode.
      parameters:
      - description: Load Flow Folder Request
        in: body
        name: request
        required: true
        schema:
          properties:
            path:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.FlowDocument'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Load flow folder
      tags:
      - flow
  /api/flow/load-path:
    post:
      consumes:
      - application/json
      description: Loads a flow document from the specified local file path. Only
        available in local mode.
      parameters:
      - description: Load Flow Request
        in: body
        name: request
        required: true
        schema:
          properties:
            path:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.FlowDocument'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Load flow from path
      tags:
      - flow
  /api/flow/open-from-system:
    post:
      consumes:
      - application/json
      description: Notifies the application that a file was opened via the operating
        system.
      parameters:
      - description: File Open Request
        in: body
        name: request
        required: true
        schema:
          properties:
            path:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Handle file open from system
      tags:
      - flow
  /api/flow/read-sources:
    post:
      consumes:
      - application/json
      description: Reads the content of specified source files.
      parameters:
      - description: Read Source Files Request
        in: body
        name: request
        required: true
        schema:
          properties:
            files:
              items:
                type: string
              type: array
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Read source files
      tags:
      - flow
  /api/flow/recent:
    get:
      description: Returns a list of recently opened flow documents.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.RecentFile'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get recent files
      tags:
      - flow
  /api/flow/remove-recent:
    post:
      consumes:
      - application/json
      description: Removes a file from the list of recently opened flow documents.
      parameters:
      - description: Remove Recent File Request
        in: body
        name: request
        required: true
        schema:
          properties:
            path:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Remove recent file
      tags:
      - flow
  /api/flow/reveal:
    post:
      consumes:
      - application/json
      description: Opens the system file manager at the specified path. Only available
        in local mode.
      parameters:
      - description: Reveal Path Request
        in: body
        name: request
        required: true
        schema:
          properties:
            path:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Reveal in file manager
      tags:
      - flow
  /api/flow/search:
    post:
      consumes:
      - application/json
      description: Performs a search for text or patterns within a specific flow document.
      parameters:
      - description: Search Flow Request
        in: body
        name: request
        required: true
        schema:
          properties:
            id:
              type: string
            query:
              $ref: '#/definitions/models.SearchQuery'
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SearchResults'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Search within flow
      tags:
      - flow
  /api/flow/source-files:
    get:
      description: Returns a list of source files related to the current flow.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              type: string
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get source files
      tags:
      - flow
...
swagger: "2.0"